# 10-10EZ | Upgrade performance on identity verification page